% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16S.R
\docType{data}
\name{data_info}
\alias{data_info}
\title{A dataframe contain information about all the dataset
\describe{
\item{ID}{Dataset name}
\item{DOI}{DOI of related paper}
\item{PRJ}{Project access number in EBI or NCBI}
\item{National}{Cohort national}
\item{Targeted}{Targeted region of the sequence}
\item{Truncated}{Parameters of truncation in dada2 pipeline}
...
}}
\format{
A dataframe contain 14 rows of dataset and 15 meta information
}
\usage{
data_info
}
\description{
A dataframe contain information about all the dataset
\describe{
\item{ID}{Dataset name}
\item{DOI}{DOI of related paper}
\item{PRJ}{Project access number in EBI or NCBI}
\item{National}{Cohort national}
\item{Targeted}{Targeted region of the sequence}
\item{Truncated}{Parameters of truncation in dada2 pipeline}
...
}
}
\keyword{datasets}
