% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16S.R
\docType{package}
\name{PD16Sdata}
\alias{PD16Sdata}
\alias{PD16Sdata-package}
\title{Eleven 16S gut microbiome Parkinson's disease dataset}
\description{
This package contains eleven 16S gut microbiome Parkinson's disease dataset.
All of the processing use dada2 pipeline(v1.38) and silva 138 taxanomy reference.
}
