% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16S.R
\docType{data}
\name{ps_list}
\alias{ps_list}
\title{Details of the dataset information can be found in ./data-raw/data_information.xlsx
\describe{
\item{ps_Hill_Burns}{A phyloseq object contain 332 samples from Hill-Burns et al.}
\item{ps_Aho}{A phyloseq object contain 266 samples from Aho et al.}
\item{ps_Vascellari}{A phyloseq object contain 114 samples from Vascellari et al.}
\item{ps_Weils}{A phyloseq object contain 256 samples from Weils et al.}
\item{ps_Heinzel}{A phyloseq object contain 663 samples from Heinzel et al.}
\item{ps_Scheperjans}{A phyloseq object contain 148 samples from Scheperjans et al.}
\item{ps_Ren}{A phyloseq object contain 39 samples from Ren et al.}
\item{ps_Jin}{A phyloseq object contain 133 samples from Jin et al.}
\item{ps_Wallen1}{A phyloseq object contain 507 samples from Wallen et al. (Dataset2 in theri paper)}
\item{ps_Wallen2}{A phyloseq object contain 333 samples from Aho et al.(Dataset1 in theri paper)}
\item{ps_Lubomski}{A phyloseq object contain 577 samples from Lubomski et al.}
}}
\format{
A list contain 11 phyloseq object
}
\usage{
ps_list
}
\description{
Details of the dataset information can be found in ./data-raw/data_information.xlsx
\describe{
\item{ps_Hill_Burns}{A phyloseq object contain 332 samples from Hill-Burns et al.}
\item{ps_Aho}{A phyloseq object contain 266 samples from Aho et al.}
\item{ps_Vascellari}{A phyloseq object contain 114 samples from Vascellari et al.}
\item{ps_Weils}{A phyloseq object contain 256 samples from Weils et al.}
\item{ps_Heinzel}{A phyloseq object contain 663 samples from Heinzel et al.}
\item{ps_Scheperjans}{A phyloseq object contain 148 samples from Scheperjans et al.}
\item{ps_Ren}{A phyloseq object contain 39 samples from Ren et al.}
\item{ps_Jin}{A phyloseq object contain 133 samples from Jin et al.}
\item{ps_Wallen1}{A phyloseq object contain 507 samples from Wallen et al. (Dataset2 in theri paper)}
\item{ps_Wallen2}{A phyloseq object contain 333 samples from Aho et al.(Dataset1 in theri paper)}
\item{ps_Lubomski}{A phyloseq object contain 577 samples from Lubomski et al.}
}
}
\keyword{datasets}
